+++ validator/db/filesystem/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "db.go",
         "genesis.go",
         "graffiti.go",
+        "import.go",
         "migration.go",
         "proposer_protection.go",
         "proposer_settings.go",
@@ -22,6 +23,8 @@ go_library(
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
         "//validator/db/common:go_default_library",
         "//validator/db/iface:go_default_library",
+        "//validator/helpers:go_default_library",
+        "//validator/slashing-protection-history/format:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
@@ -38,6 +41,7 @@ go_test(
         "db_test.go",
         "genesis_test.go",
         "graffiti_test.go",
+        "import_test.go",
         "migration_test.go",
         "proposer_protection_test.go",
         "proposer_settings_test.go",
@@ -51,8 +55,11 @@ go_test(
         "//io/file:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
+        "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//validator/db/common:go_default_library",
+        "//validator/slashing-protection-history/format:go_default_library",
+        "//validator/testing:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
     ],
