+++ validator/db/kv/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "eip_blacklisted_keys.go",
         "genesis.go",
         "graffiti.go",
+        "import.go",
         "log.go",
         "migration.go",
         "migration_optimal_attester_protection.go",
@@ -42,9 +43,13 @@ go_library(
         "//time/slots:go_default_library",
         "//validator/db/common:go_default_library",
         "//validator/db/iface:go_default_library",
+        "//validator/helpers:go_default_library",
+        "//validator/slashing-protection-history/format:go_default_library",
+        "@com_github_k0kubun_go_ansi//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
         "@com_github_prysmaticlabs_prombbolt//:go_default_library",
+        "@com_github_schollz_progressbar_v3//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_etcd_go_bbolt//:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
@@ -61,6 +66,7 @@ go_test(
         "eip_blacklisted_keys_test.go",
         "genesis_test.go",
         "graffiti_test.go",
+        "import_test.go",
         "kv_test.go",
         "migration_optimal_attester_protection_test.go",
         "migration_source_target_epochs_bucket_test.go",
@@ -82,6 +88,8 @@ go_test(
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//validator/db/common:go_default_library",
+        "//validator/slashing-protection-history/format:go_default_library",
+        "//validator/testing:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
