+++ validator/slashing-protection-history/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "doc.go",
         "export.go",
         "import.go",
-        "log.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/v5/validator/slashing-protection-history",
     visibility = [
@@ -15,22 +14,12 @@ go_library(
     ],
     deps = [
         "//config/fieldparams:go_default_library",
-        "//consensus-types/primitives:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//monitoring/progress:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
-        "//proto/prysm/v1alpha1/slashings:go_default_library",
         "//validator/db:go_default_library",
-        "//validator/db/common:go_default_library",
-        "//validator/db/filesystem:go_default_library",
-        "//validator/db/kv:go_default_library",
         "//validator/helpers:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
-        "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
-        "@com_github_k0kubun_go_ansi//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
-        "@com_github_schollz_progressbar_v3//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
 
@@ -38,7 +27,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "export_test.go",
-        "import_test.go",
         "round_trip_test.go",
     ],
     embed = [":go_default_library"],
@@ -49,10 +37,8 @@ go_test(
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//validator/db/common:go_default_library",
-        "//validator/db/kv:go_default_library",
         "//validator/db/testing:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
         "//validator/testing:go_default_library",
-        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
