+++ validator/slashing-protection-history/import.go
@@ -1,28 +1,11 @@
 package history
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
 	"io"
-	"strings"
 
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/k0kubun/go-ansi"
 	"github.com/pkg/errors"
-	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
-	"github.com/prysmaticlabs/prysm/v5/consensus-types/primitives"
-	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
-	ethpb "github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1"
-	"github.com/prysmaticlabs/prysm/v5/proto/prysm/v1alpha1/slashings"
 	"github.com/prysmaticlabs/prysm/v5/validator/db"
-	"github.com/prysmaticlabs/prysm/v5/validator/db/common"
-	"github.com/prysmaticlabs/prysm/v5/validator/db/filesystem"
-	"github.com/prysmaticlabs/prysm/v5/validator/db/kv"
-	"github.com/prysmaticlabs/prysm/v5/validator/helpers"
-	"github.com/prysmaticlabs/prysm/v5/validator/slashing-protection-history/format"
-	"github.com/schollz/progressbar/v3"
 )
 
 // ImportStandardProtectionJSON takes in EIP-3076 compliant JSON file used for slashing protection
@@ -30,513 +13,13 @@ import (
 // protection in the validator client's database. For more information, see the EIP document here:
 // https://eips.ethereum.org/EIPS/eip-3076.
 func ImportStandardProtectionJSON(ctx context.Context, validatorDB db.Database, r io.Reader) error {
-	switch validatorDB.(type) {
-	case *kv.Store:
-		return importStandardProtectionJSONComplete(ctx, validatorDB.(*kv.Store), r)
-	case *filesystem.Store:
-		return importStandardProtectionJSONMinimal(ctx, validatorDB, r)
-	default:
-		return fmt.Errorf("unsupported database type")
-	}
-}
-
-// importStandardProtectionJSON takes in EIP-3076 compliant JSON file used for slashing protection
-// by Ethereum validators and imports its data into Prysm's internal complete representation of slashing
-// protection in the validator client's database.
-func importStandardProtectionJSONComplete(ctx context.Context, validatorDB *kv.Store, r io.Reader) error {
-	encodedJSON, err := io.ReadAll(r)
-	if err != nil {
-		return errors.Wrap(err, "could not read slashing protection JSON file")
+	if validatorDB == nil {
+		return errors.New("validatorDB is nil")
 	}
 
-	interchangeJSON := &format.EIPSlashingProtectionFormat{}
-	if err := json.Unmarshal(encodedJSON, interchangeJSON); err != nil {
-		return errors.Wrap(err, "could not unmarshal slashing protection JSON file")
-	}
-
-	if interchangeJSON.Data == nil {
-		log.Warn("No slashing protection data to import")
-		return nil
-	}
-
-	// We validate the `MetadataV0` field of the slashing protection JSON file.
-	if err := helpers.ValidateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
-		return errors.Wrap(err, "slashing protection JSON metadata was incorrect")
-	}
-
-	// We need to handle duplicate public keys in the JSON file, with potentially
-	// different signing histories for both attestations and blocks.
-	signedBlocksByPubKey, err := parseBlocksForUniquePublicKeys(interchangeJSON.Data)
-	if err != nil {
-		return errors.Wrap(err, "could not parse unique entries for blocks by public key")
-	}
-
-	signedAttsByPubKey, err := parseAttestationsForUniquePublicKeys(interchangeJSON.Data)
-	if err != nil {
-		return errors.Wrap(err, "could not parse unique entries for attestations by public key")
-	}
-
-	attestingHistoryByPubKey := make(map[[fieldparams.BLSPubkeyLength]byte][]*common.AttestationRecord)
-	proposalHistoryByPubKey := make(map[[fieldparams.BLSPubkeyLength]byte]common.ProposalHistoryForPubkey)
-
-	for pubKey, signedBlocks := range signedBlocksByPubKey {
-		// Transform the processed signed blocks data from the JSON
-		// file into the internal Prysm representation of proposal history.
-		proposalHistory, err := transformSignedBlocks(ctx, signedBlocks)
-		if err != nil {
-			return errors.Wrapf(err, "could not parse signed blocks in JSON file for key %#x", pubKey)
-		}
-
-		proposalHistoryByPubKey[pubKey] = *proposalHistory
-	}
-
-	for pubKey, signedAtts := range signedAttsByPubKey {
-		// Transform the processed signed attestation data from the JSON
-		// file into the internal Prysm representation of attesting history.
-		historicalAtt, err := transformSignedAttestations(pubKey, signedAtts)
-		if err != nil {
-			return errors.Wrapf(err, "could not parse signed attestations in JSON file for key %#x", pubKey)
-		}
-
-		attestingHistoryByPubKey[pubKey] = historicalAtt
-	}
-
-	// We validate and filter out public keys parsed from JSON to ensure we are
-	// not importing those which are slashable with respect to other data within the same JSON.
-	slashableProposerKeys := filterSlashablePubKeysFromBlocks(ctx, proposalHistoryByPubKey)
-
-	slashableAttesterKeys, err := filterSlashablePubKeysFromAttestations(ctx, validatorDB, attestingHistoryByPubKey)
-	if err != nil {
-		return errors.Wrap(err, "could not filter slashable attester public keys from JSON data")
-	}
-
-	slashablePublicKeysCount := len(slashableProposerKeys) + len(slashableAttesterKeys)
-	slashablePublicKeys := make([][fieldparams.BLSPubkeyLength]byte, 0, slashablePublicKeysCount)
-	slashablePublicKeys = append(slashablePublicKeys, slashableProposerKeys...)
-	slashablePublicKeys = append(slashablePublicKeys, slashableAttesterKeys...)
-
-	if err := validatorDB.SaveEIPImportBlacklistedPublicKeys(ctx, slashablePublicKeys); err != nil {
-		return errors.Wrap(err, "could not save slashable public keys to database")
-	}
-
-	// We save the histories to disk as atomic operations, ensuring that this only occurs
-	// until after we successfully parse all data from the JSON file. If there is any error
-	// in parsing the JSON proposal and attesting histories, we will not reach this point.
-	if err := saveProposals(ctx, proposalHistoryByPubKey, validatorDB); err != nil {
-		return errors.Wrap(err, "could not save proposals")
-	}
-
-	if err := saveAttestations(ctx, attestingHistoryByPubKey, validatorDB); err != nil {
-		return errors.Wrap(err, "could not save attestations")
+	if err := validatorDB.ImportStandardProtectionJSON(ctx, r); err != nil {
+		return errors.Wrap(err, "could not import slashing protection JSON file")
 	}
 
 	return nil
-}
-
-// importStandardProtectionJSON takes in EIP-3076 compliant JSON file used for slashing protection
-// by Ethereum validators and imports its data into Prysm's internal minimal representation of slashing
-// protection in the validator client's database.
-func importStandardProtectionJSONMinimal(ctx context.Context, validatorDB db.Database, r io.Reader) error {
-	// Read the JSON file
-	encodedJSON, err := io.ReadAll(r)
-	if err != nil {
-		return errors.Wrap(err, "could not read slashing protection JSON file")
-	}
-
-	// Unmarshal the JSON file
-	interchangeJSON := &format.EIPSlashingProtectionFormat{}
-	if err := json.Unmarshal(encodedJSON, interchangeJSON); err != nil {
-		return errors.Wrap(err, "could not unmarshal slashing protection JSON file")
-	}
-
-	// If there is no data in the JSON file, we can return early.
-	if interchangeJSON.Data == nil {
-		log.Warn("No slashing protection data to import")
-		return nil
-	}
-
-	// We validate the `MetadataV0` field of the slashing protection JSON file.
-	if err := helpers.ValidateMetadata(ctx, validatorDB, interchangeJSON); err != nil {
-		return errors.Wrap(err, "slashing protection JSON metadata was incorrect")
-	}
-
-	// Save block proposals and attestations into the database
-	for _, item := range interchangeJSON.Data {
-		// If item is nil, skip
-		if item == nil {
-			continue
-		}
-
-		// Convert pubkey to bytes array
-		pubkeyBytes, err := hexutil.Decode(item.Pubkey)
-		if err != nil {
-			return errors.Wrap(err, "could not decode public key from hex")
-		}
-
-		pubkey := ([fieldparams.BLSPubkeyLength]byte)(pubkeyBytes)
-
-		// Block proposals
-		if err := importBlockProposals(ctx, pubkey, item, validatorDB); err != nil {
-			return errors.Wrap(err, "could not import block proposals")
-		}
-
-		// Attestations
-		if err := importAttestations(ctx, pubkey, item, validatorDB); err != nil {
-			return errors.Wrap(err, "could not import attestations")
-		}
-	}
-
-	return nil
-}
-
-func importBlockProposals(ctx context.Context, pubkey [fieldparams.BLSPubkeyLength]byte, item *format.ProtectionData, validatorDB db.Database) error {
-	for _, sb := range item.SignedBlocks {
-		// If signing block is nil, return early
-		if sb == nil {
-			return nil
-		}
-
-		// Convert slot to primitives.Slot
-		slot, err := helpers.SlotFromString(sb.Slot)
-		if err != nil {
-			return errors.Wrap(err, "could not convert slot to primitives.Slot")
-		}
-
-		// Save proposal if not slashable regarding EIP-3076 (minimal database)
-		if err := validatorDB.SaveProposalHistoryForSlot(ctx, pubkey, slot, []byte{}); err != nil && !strings.Contains(err.Error(), "could not sign proposal") {
-			return errors.Wrap(err, "could not save proposal history from imported JSON to database")
-		}
-	}
-
-	return nil
-}
-
-func importAttestations(ctx context.Context, pubkey [fieldparams.BLSPubkeyLength]byte, item *format.ProtectionData, validatorDB db.Database) error {
-	atts := make([]*ethpb.IndexedAttestation, len(item.SignedAttestations))
-	for i := range item.SignedAttestations {
-		// Get signed attestation
-		sa := item.SignedAttestations[i]
-
-		// Convert source epoch to primitives.Epoch
-		source, err := helpers.EpochFromString(sa.SourceEpoch)
-		if err != nil {
-			return errors.Wrap(err, "could not convert source epoch to primitives.Epoch")
-		}
-
-		// Convert target epoch to primitives.Epoch
-		target, err := helpers.EpochFromString(sa.TargetEpoch)
-		if err != nil {
-			return errors.Wrap(err, "could not convert target epoch to primitives.Epoch")
-		}
-
-		// Create indexed attestation
-		att := &ethpb.IndexedAttestation{
-			Data: &ethpb.AttestationData{
-				Source: &ethpb.Checkpoint{
-					Epoch: source,
-				},
-				Target: &ethpb.Checkpoint{
-					Epoch: target,
-				},
-			},
-		}
-
-		atts[i] = att
-	}
-
-	// Save attestations
-	if err := validatorDB.SaveAttestationsForPubKey(ctx, pubkey, [][]byte{}, atts); err != nil && !strings.Contains(err.Error(), "could not sign attestation") {
-		return errors.Wrap(err, "could not save attestation record from imported JSON to database")
-	}
-
-	return nil
-}
-
-func saveProposals(ctx context.Context, proposalHistoryByPubKey map[[fieldparams.BLSPubkeyLength]byte]common.ProposalHistoryForPubkey, validatorDB db.Database) error {
-	for pubKey, proposalHistory := range proposalHistoryByPubKey {
-		bar := initializeProgressBar(
-			len(proposalHistory.Proposals),
-			fmt.Sprintf("Importing proposals for validator public key %#x", bytesutil.Trunc(pubKey[:])),
-		)
-
-		for _, proposal := range proposalHistory.Proposals {
-			if err := bar.Add(1); err != nil {
-				log.WithError(err).Debug("Could not increase progress bar")
-			}
-
-			if err := validatorDB.SaveProposalHistoryForSlot(ctx, pubKey, proposal.Slot, proposal.SigningRoot); err != nil {
-				return errors.Wrap(err, "could not save proposal history from imported JSON to database")
-			}
-		}
-	}
-
-	return nil
-}
-
-func saveAttestations(ctx context.Context, attestingHistoryByPubKey map[[fieldparams.BLSPubkeyLength]byte][]*common.AttestationRecord, validatorDB db.Database) error {
-	bar := initializeProgressBar(
-		len(attestingHistoryByPubKey),
-		"Importing attesting history for validator public keys",
-	)
-
-	for pubKey, attestations := range attestingHistoryByPubKey {
-		if err := bar.Add(1); err != nil {
-			log.WithError(err).Debug("Could not increase progress bar")
-		}
-
-		indexedAtts := make([]*ethpb.IndexedAttestation, len(attestations))
-		signingRoots := make([][]byte, len(attestations))
-
-		for i, att := range attestations {
-			indexedAtt := createAttestation(att.Source, att.Target)
-			indexedAtts[i] = indexedAtt
-			signingRoots[i] = att.SigningRoot
-		}
-
-		if err := validatorDB.SaveAttestationsForPubKey(ctx, pubKey, signingRoots, indexedAtts); err != nil {
-			return errors.Wrap(err, "could not save attestations from imported JSON to database")
-		}
-	}
-
-	return nil
-}
-
-func initializeProgressBar(numItems int, msg string) *progressbar.ProgressBar {
-	return progressbar.NewOptions(
-		numItems,
-		progressbar.OptionFullWidth(),
-		progressbar.OptionSetWriter(ansi.NewAnsiStdout()),
-		progressbar.OptionEnableColorCodes(true),
-		progressbar.OptionSetTheme(progressbar.Theme{
-			Saucer:        "[green]=[reset]",
-			SaucerHead:    "[green]>[reset]",
-			SaucerPadding: " ",
-			BarStart:      "[",
-			BarEnd:        "]",
-		}),
-		progressbar.OptionOnCompletion(func() { fmt.Println() }),
-		progressbar.OptionSetDescription(msg),
-	)
-}
-
-// We create a map of pubKey -> []*SignedBlock. Then, for each public key we observe,
-// we append to this map. This allows us to handle valid input JSON data such as:
-//
-//	"0x2932232930: {
-//	  SignedBlocks: [Slot: 5, Slot: 6, Slot: 7],
-//	 },
-//
-//	"0x2932232930: {
-//	  SignedBlocks: [Slot: 5, Slot: 10, Slot: 11],
-//	 }
-//
-// Which should be properly parsed as:
-//
-//	"0x2932232930: {
-//	  SignedBlocks: [Slot: 5, Slot: 5, Slot: 6, Slot: 7, Slot: 10, Slot: 11],
-//	 }
-func parseBlocksForUniquePublicKeys(data []*format.ProtectionData) (map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock, error) {
-	signedBlocksByPubKey := make(map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock)
-	for _, validatorData := range data {
-		pubKey, err := helpers.PubKeyFromHex(validatorData.Pubkey)
-		if err != nil {
-			return nil, fmt.Errorf("%s is not a valid public key: %w", validatorData.Pubkey, err)
-		}
-		for _, sBlock := range validatorData.SignedBlocks {
-			if sBlock == nil {
-				continue
-			}
-			signedBlocksByPubKey[pubKey] = append(signedBlocksByPubKey[pubKey], sBlock)
-		}
-	}
-	return signedBlocksByPubKey, nil
-}
-
-// We create a map of pubKey -> []*SignedAttestation. Then, for each public key we observe,
-// we append to this map. This allows us to handle valid input JSON data such as:
-//
-//	"0x2932232930: {
-//	  SignedAttestations: [{Source: 5, Target: 6}, {Source: 6, Target: 7}],
-//	 },
-//
-//	"0x2932232930: {
-//	  SignedAttestations: [{Source: 5, Target: 6}],
-//	 }
-//
-// Which should be properly parsed as:
-//
-//	"0x2932232930: {
-//	  SignedAttestations: [{Source: 5, Target: 6}, {Source: 5, Target: 6}, {Source: 6, Target: 7}],
-//	 }
-func parseAttestationsForUniquePublicKeys(data []*format.ProtectionData) (map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedAttestation, error) {
-	signedAttestationsByPubKey := make(map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedAttestation)
-	for _, validatorData := range data {
-		pubKey, err := helpers.PubKeyFromHex(validatorData.Pubkey)
-		if err != nil {
-			return nil, fmt.Errorf("%s is not a valid public key: %w", validatorData.Pubkey, err)
-		}
-		for _, sAtt := range validatorData.SignedAttestations {
-			if sAtt == nil {
-				continue
-			}
-			signedAttestationsByPubKey[pubKey] = append(signedAttestationsByPubKey[pubKey], sAtt)
-		}
-	}
-	return signedAttestationsByPubKey, nil
-}
-
-func filterSlashablePubKeysFromBlocks(_ context.Context, historyByPubKey map[[fieldparams.BLSPubkeyLength]byte]common.ProposalHistoryForPubkey) [][fieldparams.BLSPubkeyLength]byte {
-	// Given signing roots are optional in the EIP standard, we behave as follows:
-	// For a given block:
-	//   If we have a previous block with the same slot in our history:
-	//     If signing root is nil, we consider that proposer public key as slashable
-	//     If signing root is not nil , then we compare signing roots. If they are different,
-	//     then we consider that proposer public key as slashable.
-	slashablePubKeys := make([][fieldparams.BLSPubkeyLength]byte, 0)
-	for pubKey, proposals := range historyByPubKey {
-		seenSigningRootsBySlot := make(map[primitives.Slot][]byte)
-		for _, blk := range proposals.Proposals {
-			if signingRoot, ok := seenSigningRootsBySlot[blk.Slot]; ok {
-				if signingRoot == nil || !bytes.Equal(signingRoot, blk.SigningRoot) {
-					slashablePubKeys = append(slashablePubKeys, pubKey)
-					break
-				}
-			}
-			seenSigningRootsBySlot[blk.Slot] = blk.SigningRoot
-		}
-	}
-	return slashablePubKeys
-}
-
-func filterSlashablePubKeysFromAttestations(
-	ctx context.Context,
-	validatorDB *kv.Store,
-	signedAttsByPubKey map[[fieldparams.BLSPubkeyLength]byte][]*common.AttestationRecord,
-) ([][fieldparams.BLSPubkeyLength]byte, error) {
-	slashablePubKeys := make([][fieldparams.BLSPubkeyLength]byte, 0)
-	// First we need to find attestations that are slashable with respect to other
-	// attestations within the same JSON import.
-	for pubKey, signedAtts := range signedAttsByPubKey {
-		signingRootsByTarget := make(map[primitives.Epoch][]byte)
-		targetEpochsBySource := make(map[primitives.Epoch][]primitives.Epoch)
-	Loop:
-		for _, att := range signedAtts {
-			// Check for double votes.
-			if sr, ok := signingRootsByTarget[att.Target]; ok {
-				if slashings.SigningRootsDiffer(sr, att.SigningRoot) {
-					slashablePubKeys = append(slashablePubKeys, pubKey)
-					break Loop
-				}
-			}
-			// Check for surround voting.
-			for source, targets := range targetEpochsBySource {
-				for _, target := range targets {
-					a := createAttestation(source, target)
-					b := createAttestation(att.Source, att.Target)
-					if slashings.IsSurround(a, b) || slashings.IsSurround(b, a) {
-						slashablePubKeys = append(slashablePubKeys, pubKey)
-						break Loop
-					}
-				}
-			}
-			signingRootsByTarget[att.Target] = att.SigningRoot
-			targetEpochsBySource[att.Source] = append(targetEpochsBySource[att.Source], att.Target)
-		}
-	}
-	// Then, we need to find attestations that are slashable with respect to our database.
-	for pubKey, signedAtts := range signedAttsByPubKey {
-		for _, att := range signedAtts {
-			indexedAtt := createAttestation(att.Source, att.Target)
-
-			// If slashable == NotSlashable and err != nil, then CheckSlashableAttestation failed.
-			// If slashable != NotSlashable, then err contains the reason why the attestation is slashable.
-			slashable, err := validatorDB.CheckSlashableAttestation(ctx, pubKey, att.SigningRoot, indexedAtt)
-			if err != nil && slashable == kv.NotSlashable {
-				return nil, err
-			}
-
-			if slashable != kv.NotSlashable {
-				slashablePubKeys = append(slashablePubKeys, pubKey)
-				break
-			}
-		}
-	}
-	return slashablePubKeys, nil
-}
-
-func transformSignedBlocks(_ context.Context, signedBlocks []*format.SignedBlock) (*common.ProposalHistoryForPubkey, error) {
-	proposals := make([]common.Proposal, len(signedBlocks))
-	for i, proposal := range signedBlocks {
-		slot, err := helpers.SlotFromString(proposal.Slot)
-		if err != nil {
-			return nil, fmt.Errorf("%s is not a valid slot: %w", proposal.Slot, err)
-		}
-
-		// Signing roots are optional in the standard JSON file.
-		// If the signing root is not provided, we use a default value which is a zero-length byte slice.
-		signingRoot := make([]byte, 0, fieldparams.RootLength)
-
-		if proposal.SigningRoot != "" {
-			signingRoot32, err := helpers.RootFromHex(proposal.SigningRoot)
-			if err != nil {
-				return nil, fmt.Errorf("%s is not a valid root: %w", proposal.SigningRoot, err)
-			}
-			signingRoot = signingRoot32[:]
-		}
-
-		proposals[i] = common.Proposal{
-			Slot:        slot,
-			SigningRoot: signingRoot,
-		}
-	}
-
-	return &common.ProposalHistoryForPubkey{
-		Proposals: proposals,
-	}, nil
-}
-
-func transformSignedAttestations(pubKey [fieldparams.BLSPubkeyLength]byte, atts []*format.SignedAttestation) ([]*common.AttestationRecord, error) {
-	historicalAtts := make([]*common.AttestationRecord, 0)
-	for _, attestation := range atts {
-		target, err := helpers.EpochFromString(attestation.TargetEpoch)
-		if err != nil {
-			return nil, fmt.Errorf("%s is not a valid epoch: %w", attestation.TargetEpoch, err)
-		}
-		source, err := helpers.EpochFromString(attestation.SourceEpoch)
-		if err != nil {
-			return nil, fmt.Errorf("%s is not a valid epoch: %w", attestation.SourceEpoch, err)
-		}
-
-		// Signing roots are optional in the standard JSON file.
-		// If the signing root is not provided, we use a default value which is a zero-length byte slice.
-		signingRoot := make([]byte, 0, fieldparams.RootLength)
-
-		if attestation.SigningRoot != "" {
-			signingRoot32, err := helpers.RootFromHex(attestation.SigningRoot)
-			if err != nil {
-				return nil, fmt.Errorf("%s is not a valid root: %w", attestation.SigningRoot, err)
-			}
-			signingRoot = signingRoot32[:]
-		}
-		historicalAtts = append(historicalAtts, &common.AttestationRecord{
-			PubKey:      pubKey,
-			Source:      source,
-			Target:      target,
-			SigningRoot: signingRoot,
-		})
-	}
-	return historicalAtts, nil
-}
-
-func createAttestation(source, target primitives.Epoch) *ethpb.IndexedAttestation {
-	return &ethpb.IndexedAttestation{
-		Data: &ethpb.AttestationData{
-			Source: &ethpb.Checkpoint{
-				Epoch: source,
-			},
-			Target: &ethpb.Checkpoint{
-				Epoch: target,
-			},
-		},
-	}
 }
