--- validator/slashing-protection-history/log.go
@@ -1,5 +1,0 @@
-package history
-
-import "github.com/sirupsen/logrus"
-
-var log = logrus.WithField("prefix", "slashing-protection-history")
